+++ packages/message-bundle-component/src/stories/inlang-message-bundle.styles.ts
@@ -23,6 +23,7 @@ export const messageBundleStyling = css`
 		margin-bottom: 16px;
 	}
 	.message {
+		position: relative;
 		display: flex;
 		min-height: 44px;
 		width: 100%;
@@ -44,12 +45,17 @@ export const messageBundleStyling = css`
 		display: flex;
 		flex-direction: column;
 	}
-	.selector-container {
+	.message-header {
 		width: 100%;
 		min-height: 44px;
 		display: flex;
+		justify-content: space-between;
 		background-color: var(--sl-color-neutral-100);
 		border-bottom: 1px solid var(--sl-color-neutral-300);
+	}
+	.selector-container {
+		min-height: 44px;
+		display: flex;
 	}
 	.selector {
 		height: 44px;
@@ -60,41 +66,17 @@ export const messageBundleStyling = css`
 		border-right: 1px solid var(--sl-color-neutral-300);
 		font-weight: 600;
 	}
+	.message-actions {
+		height: 44px;
+		display: flex;
+		align-items: center;
+		padding: 12px;
+	}
 	.variants-container {
 		width: 100%;
 		height: 44px;
 		display: flex;
 		flex-direction: column;
 		height: auto;
-	}
-	.variant {
-		min-height: 44px;
-		width: 100%;
-		display: flex;
-		align-items: center;
-		border-top: 1px solid var(--sl-color-neutral-300);
-	}
-	.variant:first-child {
-		border-top: none;
-	}
-	.match {
-		padding: 12px;
-		height: 44px;
-		width: 80px;
-		background-color: var(--sl-color-neutral-100);
-		border-right: 1px solid var(--sl-color-neutral-300);
-	}
-	.pattern {
-		flex: 1;
-		border: none;
-		background-color: none;
-		min-height: 20px;
-		padding: 12px;
-	}
-	.actions {
-		padding: 12px;
-		display: flex;
-		gap: 10px;
-		background-color: #fff;
 	}
 `