+++ packages/paraglide/paraglide-sveltekit/example/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @inlang/paraglide-sveltekit-example
 
+## 0.1.56
+
+### Patch Changes
+
+- @inlang/paraglide-sveltekit@0.9.4
+
+## 0.1.55
+
+### Patch Changes
+
+- Updated dependencies [def2d4e]
+  - @inlang/paraglide-sveltekit@0.9.3
+
 ## 0.1.54
 
 ### Patch Changes