+++ packages/plugins/m-function-matcher/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @inlang/plugin-m-function-matcher
 
+## 0.9.29
+
+### Patch Changes
+
+- Updated dependencies [2a5645c]
+  - @inlang/sdk@0.35.7
+
 ## 0.9.28
 
 ### Patch Changes