+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @inlang/sdk
 
+## 0.35.7
+
+### Patch Changes
+
+- 2a5645c: fix undefined lintreports for Ninja (MESDK-130)
+
 ## 0.35.6
 
 ### Patch Changes